requirements.txt
ipykernel
matplotlib
seaborn
scikit-learn
xgboost
torch
torchvision
Flask
flask-mongoengine
flask-jwt-extended
python-dotenv
bcrypt
pymongo
joblib
tensorflow
numpy
requests
langdetect
fastapi
torchvision
torch
torchaudio
flask-cors
uvicorn
python-multipart
groq